backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1

# Vector DB
pinecone==4.1.0

# LLM
groq==0.9.0

# Embeddings
sentence-transformers==3.0.1
torch==2.3.0
torchvision==0.18.0

# PDF processing
pdfplumber==0.11.1

# Image OCR
Pillow==10.3.0
pytesseract==0.3.10

# Audio transcription
openai-whisper==20231117
ffmpeg-python==0.2.0

# HTTP / utils
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
